Configurable file hasher to allow using other implementations
