+++ src/config.ts
@@ -1,5 +1,14 @@
 import {DummyLogger} from './util/DummyLogger';
-import {Logger} from './util/types';
+import {SHA1FileHasher} from './util/SHA1FileHasher';
+import {FileHasher, Logger} from './util/types';
+
+let globalFileHasher = new SHA1FileHasher();
+export function setFileHasher(fileHasher: FileHasher): void {
+    globalFileHasher = fileHasher;
+}
+export function getFileHasher(): FileHasher {
+    return globalFileHasher;
+}
 
 let globalLogger = new DummyLogger();
 export function setLogger(logger: Logger): void {
