+++ src/util/types.ts
@@ -1,3 +1,7 @@
+export interface FileHasher {
+    hash(file: string): Promise<string>
+}
+
 export interface Logger {
     debug(msg: string): void
     info(msg: string): void
